requirements.txt
numpy>=1.21.4
Pillow>=9.5.0
torch>=2.0.0
